Update login and register forms, separate interfaces and fetch requests from components
