--- todo-list-enh.client/src/components/Interfaces/User.ts
@@ -1,6 +1,0 @@
-export default interface User {
-    id: number;
-    username: string;
-    email: string;
-    createdAt: string;
-}
