+++ todo-list-enh.client/src/components/pages/Register.tsx
@@ -4,9 +4,9 @@ import { yupResolver } from "@hookform/resolvers/yup";
 import * as yup from "yup";
 import { Button, Container, Typography, Box, Alert } from "@mui/material";
 import { register } from "../reusable-items/AuthController";
+import { AuthRequest } from "../Interfaces/UserInterfaces";
 import FormInput from "../reusable-items/FormInput";
 
-// Validation schema
 const schema = yup.object().shape({
     username: yup.string().min(3, "Username must be at least 3 characters").required("Username is required"),
     email: yup.string().email("Invalid email format").required("Email is required"),
@@ -19,7 +19,7 @@ export default function Register() {
         resolver: yupResolver(schema),
     });
 
-    const onSubmit = async (data: any) => {
+    const onSubmit = async (data: AuthRequest) => {
         try {
             const result = await register(data);
             localStorage.setItem("token", result.token);
