+++ todo-list-enh.client/src/components/reusable-items/AuthController.ts
@@ -1,35 +1,25 @@
-export interface User {
-    id: number;
-    username: string;
-    email: string;
-    createdAt: string;
-}
-export interface AuthResponse {
-    token: string;
-    user: User;
-}
-
-export interface AuthRequest {
-    email: string;
-    password: string;
-    user?: User;
-}
+import { AuthRequest, AuthResponse } from "../Interfaces/UserInterfaces"
 
 const API_URL = "https://localhost:7289/Users";
 
-export const login = async (data: AuthRequest): Promise<AuthResponse> => {
+export async function loginUser(email : string, password : string) {
     const response = await fetch(`${API_URL}/login`, {
         method: "POST",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(data),
+        body: JSON.stringify({
+            username: "undefined",
+            email,
+            password,
+            createdAt: new Date().toISOString(),
+        }),
     });
 
     if (!response.ok) {
         throw new Error("Wrong email or password.");
     }
 
-    return response.json();
-};
+    return await response.json();
+}
 
 export const register = async (data: AuthRequest): Promise<AuthResponse> => {
     const response = await fetch(`${API_URL}/register`, {
