+++ todo-list-enh.client/todo-list-enh.client.esproj
@@ -8,7 +8,4 @@
     <!-- Folder where production build objects will be placed -->
     <BuildOutputFolder>$(MSBuildProjectDirectory)\dist</BuildOutputFolder>
   </PropertyGroup>
-  <ItemGroup>
-    <Folder Include="src\components\Interfaces\" />
-  </ItemGroup>
 </Project>
